feat: add `useResponsiveImageView` hook (#37)
* feat: add `useResponsiveImageView` hook

This exposes the functionality of `react-native-responsive-image-view` as a custom hook (while
retaining the existing render prop component). In fact, all of the logic is now implemented in the
hook and the component is just a wrapper around it.

BREAKING CHANGE: `react-native@0.59.0` and `react@16.8.3` are the new minimum versions required.
Other than that, there are no migration steps necessary - your existing code will continue to work!
See https://facebook.github.io/react-native/blog/2019/03/12/releasing-react-native-059 for RN 0.59
release details. If you are unable to upgrade, you should continue using the latest 1.x release of
this library.